create BassDrumRow component and transfer code from App component to it
