+++ src/BassDrumRow/BassDrumRow.css
@@ -1,0 +1,20 @@
+.track {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+
+.note {
+    border: 1px solid black;
+    height: 100px;
+    width: 100px;
+    cursor: pointer;
+  }
+  
+  .red {
+    background-color: red;
+  }
+  
+  .green {
+    background-color: green;
+  }
